Move dashboard-click-drill into visualizations-click-ations
