+++ frontend/src/metabase/visualizations/click-actions/modes/EmbeddingSdkMode.ts
@@ -1,8 +1,3 @@
-import {
-  getClickBehavior,
-  getClickBehaviorData,
-  getParameterValuesBySlug,
-} from "metabase/dashboard/utils/dashboard-click-drill";
 import type { ParameterValues } from "metabase/embedding-sdk/types/dashboard";
 
 import type {
@@ -15,6 +10,11 @@ import { DashboardClickAction } from "../actions/DashboardClickAction";
 import { ExtractColumnAction } from "../actions/ExtractColumnAction";
 import { HideColumnAction } from "../actions/HideColumnAction";
 import { NativeQueryClickFallback } from "../actions/NativeQueryClickFallback";
+import {
+  getClickBehavior,
+  getClickBehaviorData,
+  getParameterValuesBySlug,
+} from "../lib/dashboard-click-drill";
 
 export type ClickBehaviorTarget = {
   type: "dashboard" | "question";
